move stream into pten
