+++ paddle/fluid/memory/malloc.cc
@@ -16,7 +16,7 @@ limitations under the License. */
 
 #include "paddle/fluid/memory/allocation/allocator_facade.h"
 #include "paddle/fluid/platform/place.h"
-#include "paddle/fluid/platform/stream/stream.h"
+#include "paddle/pten/core/stream.h"
 
 namespace paddle {
 namespace memory {
@@ -36,13 +36,13 @@ uint64_t Release(const platform::Place& place) {
 
 std::shared_ptr<Allocation> AllocShared(const platform::Place& place,
                                         size_t size,
-                                        const platform::Stream& stream) {
+                                        const pten::Stream& stream) {
   return allocation::AllocatorFacade::Instance().AllocShared(place, size,
                                                              stream);
 }
 
 bool InSameStream(const std::shared_ptr<Allocation>& allocation,
-                  const platform::Stream& stream) {
+                  const pten::Stream& stream) {
   return allocation::AllocatorFacade::Instance().InSameStream(allocation,
                                                               stream);
 }
